Added input system

+++ Catalyst/catalyst-engine/source/Engine/Time.cpp
@@ -4,20 +4,20 @@
 
 namespace Catalyst
 {
-	double Time::m_currTime;
-	double Time::m_prevTime;
+	float Time::m_currTime;
+	float Time::m_prevTime;
 
-	double Time::m_deltaTime;
+	float Time::m_deltaTime;
 	uint32 Time::m_frames;
-	double Time::m_fpsInterval;
+	float Time::m_fpsInterval;
 	uint32 Time::m_fps;
 
-	double Time::DeltaTime()
+	float Time::DeltaTime()
 	{
 		return m_deltaTime;
 	}
 
-	double Time::AppTime()
+	float Time::AppTime()
 	{
 		return m_currTime;
 	}
@@ -29,7 +29,7 @@ namespace Catalyst
 
 	void Time::Init()
 	{
-		m_prevTime = glfwGetTime();
+		m_prevTime = static_cast<float>(glfwGetTime());
 		m_currTime = 0;
 		m_deltaTime = 0;
 		m_frames = 0;
@@ -39,7 +39,7 @@ namespace Catalyst
 	bool Time::Tick()
 	{
 		// Update delta time
-		m_currTime = glfwGetTime();
+		m_currTime = static_cast<float>(glfwGetTime());
 		m_deltaTime = m_currTime - m_prevTime;
 		if (m_deltaTime > 0.1f)
 			m_deltaTime = 0.1f;
